Separate authentication routes from home route
