+++ src/controllers/homeController.js
@@ -8,25 +8,7 @@ const homepage = (req, res) => {
 	res.render("index", context);
 };
 
-// [GET] Login page
-const login = (req, res) => {
-	let context = {
-		title: "Login",
-	};
-	res.render("auth/login", context);
-};
-
-// [GET] Register page
-const register = (req, res) => {
-	let context = {
-		title: "Register",
-	};
-	res.render("auth/register", context);
-};
-
 // exports
 module.exports = {
-	homepage,
-	login,
-	register,
+	homepage
 };
