+++ src/controllers/index.js
@@ -1,7 +1,9 @@
 // modules
 const homeController = require("../controllers/homeController");
+const authController = require("../controllers/authController");
 
 // exports
 module.exports = {
 	homeController,
+	authController
 };
