+++ src/routes/index.js
@@ -1,7 +1,9 @@
 // modules
 const homeRoutes = require("./homeRoutes");
+const authRoutes = require("./authRoutes");
 
 // exports
 module.exports = {
 	homeRoutes,
+	authRoutes
 };
