+++ src/views/index.ejs
@@ -9,8 +9,8 @@
     
     <!-- Content -->
     <main>
-        <a href="/login">Login</a>
-        <a href="/register">Signup</a>
+        <a href="/auth/login">Login</a>
+        <a href="/auth/register">Signup</a>
     </main>
     
     <!-- Sidebar -->
